Move cycle setting getters out of the operator module and into utils
Also move the associated tests
