+++ armi/utils/__init__.py
@@ -76,6 +76,246 @@ def copyWithoutBlocking(src, dest):
     return t
 
 
+def getPowerFractions(cs):
+    """
+    Return the power fractions for each cycle.
+
+    Parameters
+    ----------
+    cs : case settings object
+
+    Returns
+    -------
+    powerFractions : 2-list
+        A list with nCycles elements, where each element is itself a list of the
+        power fractions at each step of the cycle.
+
+    Notes
+    -----
+    This is stored outside of the Operator class so that it can be easily called
+    to resolve case settings objects in other contexts (i.e. in the preparation
+    of restart runs).
+    """
+    if cs["cycles"] != []:
+        return [
+            expandRepeatedFloats(cycle["power fractions"]) for cycle in cs["cycles"]
+        ]
+    else:
+        valuePerCycle = (
+            expandRepeatedFloats(cs["powerFractions"])
+            if cs["powerFractions"] not in [None, []]
+            else [1.0] * cs["nCycles"]
+        )
+
+        return [[value] * cs["burnSteps"] for value in valuePerCycle]
+
+
+def getCycleNames(cs):
+    """
+    Return the names of each cycle. If a name is omitted, it is `None`.
+
+    Parameters
+    ----------
+    cs : case settings object
+
+    Returns
+    -------
+    cycleNames : list
+        A list of the availability factors.
+
+    Notes
+    -----
+    This is stored outside of the Operator class so that it can be easily called
+    to resolve case settings objects in other contexts (i.e. in the preparation
+    of restart runs).
+    """
+    if cs["cycles"] != []:
+        return [
+            (cycle["name"] if "name" in cycle.keys() else None)
+            for cycle in cs["cycles"]
+        ]
+    else:
+        return [None] * cs["nCycles"]
+
+
+def getAvailabilityFactors(cs):
+    """
+    Return the availability factors for each cycle.
+
+    Parameters
+    ----------
+    cs : case settings object
+
+    Returns
+    -------
+    availabilityFactors : list
+        A list of the availability factors.
+
+    Notes
+    -----
+    This is stored outside of the Operator class so that it can be easily called
+    to resolve case settings objects in other contexts (i.e. in the preparation
+    of restart runs).
+    """
+    if cs["cycles"] != []:
+        availabilityFactors = []
+        for cycle in cs["cycles"]:
+            if "availability factor" in cycle.keys():
+                availabilityFactors.append(cycle["availability factor"])
+            else:
+                availabilityFactors.append(1)
+        return availabilityFactors
+    else:
+        return (
+            expandRepeatedFloats(cs["availabilityFactors"])
+            if cs["availabilityFactors"] not in [None, []]
+            else (
+                [cs["availabilityFactor"]] * cs["nCycles"]
+                if cs["availabilityFactor"] != None
+                else [1]
+            )
+        )
+
+
+def _getStepAndCycleLengths(cs):
+    """
+    These need to be gotten together because it is a chicken/egg depending on which
+    style of cycles input the user employs.
+
+    Note that using this method directly is more effecient than calling `getStepLengths`
+    and `getCycleLengths` separately, but it is probably more clear to the user
+    to call each of them separately.
+    """
+    stepLengths = []
+    availabilityFactors = getAvailabilityFactors(cs)
+    if cs["cycles"] != []:
+        for cycleIdx, cycle in enumerate(cs["cycles"]):
+            cycleKeys = cycle.keys()
+
+            if "step days" in cycleKeys:
+                stepLengths.append(expandRepeatedFloats(cycle["step days"]))
+            elif "cumulative days" in cycleKeys:
+                cumulativeDays = cycle["cumulative days"]
+                stepLengths.append(getStepsFromValues(cumulativeDays))
+            elif "burn steps" in cycleKeys and "cycle length" in cycleKeys:
+                stepLengths.append(
+                    [
+                        cycle["cycle length"]
+                        * availabilityFactors[cycleIdx]
+                        / cycle["burn steps"]
+                    ]
+                    * cycle["burn steps"]
+                )
+            else:
+                raise ValueError(
+                    f"No cycle time history is given in the detailed cycles history for cycle {cycleIdx}"
+                )
+
+        cycleLengths = [sum(cycleStepLengths) for cycleStepLengths in stepLengths]
+        cycleLengths = [
+            cycleLength / aFactor
+            for (cycleLength, aFactor) in zip(cycleLengths, availabilityFactors)
+        ]
+
+    else:
+        cycleLengths = (
+            expandRepeatedFloats(cs["cycleLengths"])
+            if cs["cycleLengths"] not in [None, []]
+            else (
+                [cs["cycleLength"]] * cs["nCycles"]
+                if cs["cycleLength"] != None
+                else [0]
+            )
+        )
+        cycleLengthsModifiedByAvailability = [
+            length * availability
+            for (length, availability) in zip(cycleLengths, availabilityFactors)
+        ]
+        stepLengths = (
+            [
+                [length / cs["burnSteps"]] * cs["burnSteps"]
+                for length in cycleLengthsModifiedByAvailability
+            ]
+            if cs["burnSteps"] not in [0, None]
+            else [[]]
+        )
+
+    return stepLengths, cycleLengths
+
+
+def getStepLengths(cs):
+    """
+    Return the length of each step in each cycle.
+
+    Parameters
+    ----------
+    cs : case settings object
+
+    Returns
+    -------
+    stepLengths : 2-list
+        A list with elements for each cycle, where each element itself is a list
+        containing the step lengths in days.
+
+    Notes
+    -----
+    This is stored outside of the Operator class so that it can be easily called
+    to resolve case settings objects in other contexts (i.e. in the preparation
+    of restart runs).
+    """
+    return _getStepAndCycleLengths(cs)[0]
+
+
+def getCycleLengths(cs):
+    """
+    Return the lengths of each cycle in days.
+
+    Parameters
+    ----------
+    cs : case settings object
+
+    Returns
+    -------
+    cycleLengths : list
+        A list of the cycle lengths in days.
+
+    Notes
+    -----
+    This is stored outside of the Operator class so that it can be easily called
+    to resolve case settings objects in other contexts (i.e. in the preparation
+    of restart runs).
+    """
+    return _getStepAndCycleLengths(cs)[1]
+
+
+def getBurnSteps(cs):
+    """
+    Return the number of burn steps for each cycle.
+
+    Parameters
+    ----------
+    cs : case settings object
+
+    Returns
+    -------
+    burnSteps : list
+        A list of the number of burn steps.
+
+    Notes
+    -----
+    This is stored outside of the Operator class so that it can be easily called
+    to resolve case settings objects in other contexts (i.e. in the preparation
+    of restart runs).
+    """
+    stepLengths = getStepLengths(cs)
+    return [len(steps) for steps in stepLengths]
+
+
+def getMaxBurnSteps(cs):
+    burnSteps = getBurnSteps(cs)
+    return max(burnSteps)
+
+
 def getTimeStepNum(cycleNumber, subcycleNumber, cs):
     """Return the timestep associated with cycle and tn.
